clean up

+++ script/HelperConfig.s.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.19;
 
-import {Script} from 'forge-std/Script.sol';
+import {Script} from "forge-std/Script.sol";
 import {VRFCoordinatorV2_5Mock} from "@chainlink/src/v0.8/vrf/mocks/VRFCoordinatorV2_5Mock.sol";
-import {LinkToken} from "../test/mocks/LinkToken.sol";
+import {LinkToken} from "test/mocks/LinkToken.sol";
 
 abstract contract CodeConstances {
     /** VRG Mock Values */
@@ -24,7 +24,7 @@ contract HelperConfig is Script, CodeConstances {
         uint256 interval;
         address vrfCoordinator;
         bytes32 gasLane;
-        uint64 subscriptionId;
+        uint256 subscriptionId;
         uint32 callbackGasLimit;
         address link;
     }
@@ -78,7 +78,7 @@ contract HelperConfig is Script, CodeConstances {
         localNetworkConfig = NetworkConfig({
             entranceFee: 0.01 ether,
             interval: 30,
-            vrfCoordinator: address(vrfCoordinatorMock), // https://docs.chain.link/vrf/v2-5/supported-networks#sepolia-testnet
+            vrfCoordinator: address(vrfCoordinatorMock),
             gasLane: 0x787d74caea10b2b357790d5b5247c2f63d1d91572a9846f780606e4d953677ae,
             callbackGasLimit: 500000,
             subscriptionId: 0,
